+++ src/gcovr/formats/gcov/parser/text.py
@@ -36,7 +36,6 @@ The behavior of this parser was informed by the following sources:
 # pylint: disable=too-many-lines
 
 import enum
-import logging
 import re
 from typing import (
     Any,
@@ -47,15 +46,15 @@ from typing import (
     Union,
 )
 
+from ....data_model.coverage import FileCoverage, LineCoverage
+from ....data_model.merging import FUNCTION_MAX_LINE_MERGE_OPTIONS, MergeOptions
+from ....logging import LOGGER
+from ....utils import get_md5_hexdigest
+
 from .common import (
     SUSPICIOUS_COUNTER,
     check_hits,
 )
-from ....utils import get_md5_hexdigest
-from ....data_model.coverage import FileCoverage, LineCoverage
-from ....data_model.merging import FUNCTION_MAX_LINE_MERGE_OPTIONS, MergeOptions
-
-LOGGER = logging.getLogger("gcovr")
 
 
 def _line_pattern(pattern: str) -> Pattern[str]:
@@ -204,6 +203,7 @@ def parse_metadata(
     lines: list[str],
     *,
     suspicious_hits_threshold: int = SUSPICIOUS_COUNTER,
+    trace_file: bool = False,
 ) -> dict[str, Optional[str]]:
     r"""
     Collect the header/metadata lines from a gcov file.
@@ -237,6 +237,8 @@ def parse_metadata(
 
         if isinstance(parsed_line, _MetadataLine):
             key, value = parsed_line
+            if trace_file:
+                LOGGER.trace(f"Metadata: {parsed_line.key}={parsed_line.value}")
             collected[key] = value
         else:
             break  # stop at the first line that is not metadata
@@ -260,6 +262,7 @@ def parse_coverage(
     filename: str,
     ignore_parse_errors: Optional[set[str]],
     suspicious_hits_threshold: int = SUSPICIOUS_COUNTER,
+    trace_file: bool = False,
 ) -> tuple[FileCoverage, list[str]]:
     """
     Extract coverage data from a gcov report.
@@ -289,18 +292,16 @@ def parse_coverage(
             continue
 
         try:
-            tokenized_lines.append(
-                (
-                    _parse_line(
-                        filename,
-                        raw_line,
-                        suspicious_hits_threshold,
-                        ignore_parse_errors,
-                        persistent_states,
-                    ),
-                    raw_line,
-                )
+            parsed_line = _parse_line(
+                filename,
+                raw_line,
+                suspicious_hits_threshold,
+                ignore_parse_errors,
+                persistent_states,
             )
+            tokenized_lines.append((parsed_line, raw_line))
+            if trace_file:
+                LOGGER.trace(f"Parsed line: {parsed_line}")
         except Exception as ex:  # pylint: disable=broad-except
             lines_with_errors.append((raw_line, ex))