+++ src/gcovr/formats/txt/__init__.py
@@ -17,7 +17,6 @@
 #
 # ****************************************************************************
 
-import logging
 from typing import Union
 
 from ...data_model.container import CoverageContainer
@@ -27,8 +26,6 @@ from ...options import (
     GcovrDeprecatedConfigOptionAction,
     OutputOrDefault,
 )
-
-LOGGER = logging.getLogger("gcovr")
 
 
 class UseBranchMetricAction(GcovrDeprecatedConfigOptionAction):