+++ tests/bazel/test_bazel.py
@@ -1,5 +1,4 @@
 import os
-from pathlib import Path
 import shutil
 
 import pytest
@@ -9,20 +8,11 @@ from tests.conftest import IS_DARWIN, IS_GCC, IS_WINDOWS, GcovrTestExec
 BAZEL = shutil.which("bazel")
 
 
-@pytest.mark.skipif(
-    BAZEL is None or IS_WINDOWS or (IS_DARWIN and IS_GCC) or not IS_GCC,
-    reason="Bazel test not working on Windows or on MacOs (with gcc).",
-)
-def test(gcovr_test_exec: "GcovrTestExec") -> None:
-    """Test bazel build."""
+def build_bazel_example(gcovr_test_exec: "GcovrTestExec") -> dict[str, str]:
+    """Build the bazel example to verify that bazel is working."""
     bazel_build_options = [
-        "--collect_code_coverage=True",
-        "--test_output=all",
-        "--test_env=VERBOSE_COVERAGE=1",
-    ]
-    bazel_coverage_options = [
         "--instrumentation_filter=//:lib",
-        "--experimental_fetch_all_coverage_outputs",
+        "--collect_code_coverage=True",
         "--test_output=all",
         "--test_env=VERBOSE_COVERAGE=1",
     ]
@@ -30,7 +20,6 @@ def test(gcovr_test_exec: "GcovrTestExec") -> None:
         bazel_build_options.append("--force_pic")
     if gcovr_test_exec.is_llvm():
         bazel_build_options.append("--config=clang-gcov")
-        bazel_coverage_options.append("--config=clang-gcov")
 
     env = os.environ.copy()
     env.update(
@@ -48,17 +37,24 @@ def test(gcovr_test_exec: "GcovrTestExec") -> None:
         env=env,
     )
 
-    def remove_gcda(directory: Path) -> None:
-        for file in directory.resolve().rglob("*.gcda"):
+    # Remove all existing gcda files
+    for directory in (gcovr_test_exec.output_dir / "bazel-out").glob("*-fastbuild"):
+        for file in directory.rglob("*.gcda"):
             file.unlink()
 
-    # Remove all gcda files in the bin directory
-    for directory in [
-        p / "bin"
-        for p in (gcovr_test_exec.output_dir / "bazel-out").glob("*-fastbuild")
-    ]:
-        remove_gcda(directory)
-        gcovr_test_exec.run(directory / "test" / "testcase")
+    return env
+
+
+@pytest.mark.skipif(
+    BAZEL is None or IS_WINDOWS or (IS_DARWIN and IS_GCC) or not IS_GCC,
+    reason="Bazel test not working on Windows or on MacOs (with gcc).",
+)
+def test_run_on_our_own(gcovr_test_exec: "GcovrTestExec") -> None:
+    """Test bazel build."""
+    build_bazel_example(gcovr_test_exec)
+
+    for directory in (gcovr_test_exec.output_dir / "bazel-out").glob("*-fastbuild"):
+        gcovr_test_exec.run(directory / "bin" / "test" / "testcase")
         additional_options = []
         if gcovr_test_exec.use_gcc_json_format():
             additional_options += ["--root", "/proc/self/cwd"]
@@ -67,32 +63,58 @@ def test(gcovr_test_exec: "GcovrTestExec") -> None:
             "--json-pretty",
             "--json=coverage.json",
             *additional_options,
-            directory,
+            directory / "bin",
+        )
+        break
+    else:
+        raise AssertionError("Can't resolve directory bazel-out/*-fastbuild.")
+
+    gcovr_test_exec.compare_json()
+
+
+@pytest.mark.skipif(
+    BAZEL is None or IS_WINDOWS or IS_DARWIN or not IS_GCC,
+    reason="Bazel test not working on Windows and MacOs or with LLVM/clang.",
+)
+def test_bazel_coverage(gcovr_test_exec: "GcovrTestExec") -> None:
+    """Test bazel build."""
+    env = build_bazel_example(gcovr_test_exec)
+
+    for directory in (gcovr_test_exec.output_dir / "bazel-out").glob("*-fastbuild"):
+        bazel_coverage_options = [
+            "--instrumentation_filter=//:lib",
+            "--experimental_fetch_all_coverage_outputs",
+            "--test_output=all",
+            "--test_env=VERBOSE_COVERAGE=1",
+        ]
+        if not gcovr_test_exec.is_windows():
+            bazel_coverage_options.append("--force_pic")
+        gcovr_test_exec.run(
+            str(BAZEL),
+            "coverage",
+            *bazel_coverage_options,
+            "//test:testcase",
+            env=env,
         )
 
-        if gcovr_test_exec.is_darwin():
-            (gcovr_test_exec.output_dir / "coverage_bazel.json").write_text(
-                '"Test not working"\n',
-                encoding="utf-8",
-            )
-        else:
-            remove_gcda(directory)
-            gcovr_test_exec.run(
-                str(BAZEL),
-                "coverage",
-                *bazel_coverage_options,
-                "//test:testcase",
-                env=env,
-            )
-            gcovr_test_exec.gcovr(
-                "--json-pretty",
-                "--json=coverage_bazel.json",
-                *additional_options,
-                directory.parent / "testlogs",
-            )
+        additional_options = []
+        if gcovr_test_exec.use_gcc_json_format():
+            additional_options += [
+                "--root",
+                "/proc/self/cwd",
+                "--gcov-use-existing",
+                "--gcov-keep",
+            ]
+        gcovr_test_exec.gcovr(
+            "--verbose",
+            "--json-pretty",
+            "--json=coverage.json",
+            *additional_options,
+            directory / "testlogs",
+        )
 
         break
     else:
-        raise AssertionError("Can't resolve directory bazel-out/*-fastbuild/bin.")
+        raise AssertionError("Can't resolve directory bazel-out/*-fastbuild.")
 
     gcovr_test_exec.compare_json()