+++ CHANGELOG.rst
@@ -37,6 +37,8 @@ Internal changes:
   - Remove test with MacOs 13 because image will be retired by December 4th, 2025.
 
 - Activate parallel test execution by default and improve IDE support. (:issue:`1174`)
+- Add :option:`--trace-include` and :option:`--trace-exclude` to get additional output
+  from parsing of included gcov files. (:issue:`1175`)
 - Add ``gcc-7``, ``gcc-15``, ``clang-11``, ``clang-12`` and ``clang-20`` to the test suite. (:issue:`1177`)
 - Use pyenv to install python independent from OS version in docker image for testing. (:issue:`1181`)